+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- promtail-config reconciler: creates promtail-config as extra-values.
+
 ### Changed
 
 - Push app to aws-app-catalog